+++ packages/classes/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Package now specifies Node 16.x or above.
+- AvroPayload now uses `long` instead of `int` for duration and position.
 
 ## [2.0.0-alpha.1] - 2021-04-16
 ### Changed